+++ client/app/src/main/java/com/example/witt/data/api/SignInService.kt
@@ -1,10 +1,10 @@
 package com.example.witt.data.api
 
-import com.example.witt.data.model.auth.request.SignInRequest
-import com.example.witt.data.model.auth.request.SocialSignInRequest
-import com.example.witt.data.model.auth.response.SignInResponse
-import com.example.witt.data.model.auth.response.SocialSignInResponse
-import com.example.witt.data.model.user.response.TokenResponse
+import com.example.witt.data.model.remote.signin.request.SignInRequest
+import com.example.witt.data.model.remote.signin.request.SocialSignInRequest
+import com.example.witt.data.model.remote.signin.response.SignInResponse
+import com.example.witt.data.model.remote.signin.response.SocialSignInResponse
+import com.example.witt.data.model.remote.user.response.TokenResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
