+++ client/app/src/main/java/com/example/witt/data/model/remote/signin/request/SignInRequest.kt
@@ -1,4 +1,4 @@
-package com.example.witt.data.model.auth.request
+package com.example.witt.data.model.remote.signin.request
 
 data class SignInRequest(
     val AccountType : Int,
