+++ client/app/src/main/java/com/example/witt/data/model/remote/signin/response/SocialSignInResponse.kt
@@ -1,6 +1,6 @@
-package com.example.witt.data.model.auth.response
+package com.example.witt.data.model.remote.signin.response
 
-data class SignInResponse (
+data class SocialSignInResponse(
     val status : Boolean,
     val reason: String,
     val AccessToken : String,
