+++ client/app/src/main/java/com/example/witt/data/source/remote/signin/SignInDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.witt.data.source.remote.signin
 
-import com.example.witt.data.model.auth.request.SignInRequest
-import com.example.witt.data.model.auth.response.SignInResponse
+import com.example.witt.data.model.remote.signin.request.SignInRequest
+import com.example.witt.data.model.remote.signin.response.SignInResponse
 
 interface SignInDataSource {
     suspend fun signIn(signInRequest: SignInRequest) : Result<SignInResponse>
