--- client/app/src/main/java/com/example/witt/data/source/remote/user/TokenSignInDataSource.kt
@@ -1,9 +1,0 @@
-package com.example.witt.data.source.remote.token_signin
-
-import com.example.witt.data.model.user.response.TokenResponse
-
-
-interface TokenSignInDataSource {
-
-    suspend fun tokenSignIn() : Result<TokenResponse>
-}
