+++ client/app/src/main/java/com/example/witt/domain/model/remote/signin/SignInModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.auth
+package com.example.witt.domain.model.remote.signin
 
 data class SignInModel (
     val status: Boolean,
