+++ client/app/src/main/java/com/example/witt/domain/model/remote/signin/SocialSignInModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.auth
+package com.example.witt.domain.model.remote.signin
 
 data class SocialSignInModel (
         val status: Boolean,
