+++ client/app/src/main/java/com/example/witt/domain/model/remote/signin/TokenModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.auth
+package com.example.witt.domain.model.remote.signin
 
 data class TokenModel(
     val status: Boolean,
