+++ client/app/src/main/java/com/example/witt/domain/use_case/remote/TokenSignIn.kt
@@ -1,6 +1,6 @@
 package com.example.witt.domain.use_case.remote
 
-import com.example.witt.domain.model.auth.TokenModel
+import com.example.witt.domain.model.remote.signin.TokenModel
 import com.example.witt.domain.repository.AuthRepository
 import javax.inject.Inject
